[DI] Turn private defs to non-public ones before removing passes
