+++ src/Symfony/Component/DependencyInjection/Compiler/PassConfig.php
@@ -71,6 +71,12 @@ class PassConfig
             new CheckArgumentsValidityPass(false),
         ));
 
+        $this->beforeRemovingPasses = array(
+            -100 => array(
+                new ResolvePrivatesPass(),
+            ),
+        );
+
         $this->removingPasses = array(array(
             new RemovePrivateAliasesPass(),
             new ReplaceAliasByActualDefinitionPass(),
